pyproject.toml
[project]
name = "stream"
version = "0.1.0"
description = "Stream is a modular pipeline designed to generate synthetic or use real hospital data, format it for instruction tuning, and fine-tune a large language model (LLM) using this data."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bert-score>=0.3.13",
    "evaluate>=0.4.3",
    "mlflow>=2.22.0",
    "numpy>=2.0.2",
    "ollama>=0.4.8",
    "pandas>=2.2.3",
    "polars>=1.29.0",
    "pydantic>=2.11.3",
    "pynvml>=12.0.0",
    "pyyaml>=6.0.2",
    "torch>=2.6.0",
    "unsloth>=2024.11.10",
]
authors = [
    {name = "Fuchs Basile", email = "basile.fuchs@chu-brest.fr"},
    {name= "Lamard Arthur", email = "arthur.lamard@chu-brest.fr"}
]
maintainers = [
    {name = "Fuchs Basile", email = "basile.fuchs@chu-brest.fr"},
    {name= "Lamard Arthur", email = "arthur.lamard@chu-brest.fr"}
]
license = "GLP-3.0"
